Backend/requirements.txt
flask
flask-cors
gunicorn
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
timm
opencv-python-headless
python-dotenv
groq
numpy
requests